frameworks/langchain/requirements.txt
mlx-lm
python-dotenv
langchain-core
llama-cpp-python 
langchain
langchain-community
fastapi
uvicorn
anthropic
mcp
httpx[socks]